svelte: remove prop type export
