--- packages/dialog/svelte/lib/DialogClose.ts
@@ -1,3 +1,0 @@
-export type DialogCloseProps = svelteHTML.IntrinsicElements['button'] & {
-	node?: HTMLButtonElement | undefined | null;
-};
