+++ packages/dialog/svelte/lib/DialogContent.svelte
@@ -1,10 +1,15 @@
+<script lang="ts" context="module">
+	export type DialogContentProps = svelteHTML.IntrinsicElements['div'] & {
+		node?: HTMLDivElement | undefined | null;
+	};
+</script>
+
 <script lang="ts">
 	import {DialogContentModel} from '@ally-ui/core-dialog';
 	import {createEventForwarder} from '@ally-ui/svelte';
 	import {get_current_component, onMount} from 'svelte/internal';
 	import {readable} from 'svelte/store';
 	import {getDialogRootModel, getDialogRootState} from './context';
-	import type {DialogContentProps} from './DialogContent';
 
 	type $$Props = DialogContentProps;
