--- packages/dialog/svelte/lib/DialogContent.ts
@@ -1,3 +1,0 @@
-export type DialogContentProps = svelteHTML.IntrinsicElements['div'] & {
-	node?: HTMLDivElement | undefined | null;
-};
