--- packages/dialog/svelte/lib/DialogDescription.ts
@@ -1,3 +1,0 @@
-export type DialogDescriptionProps = svelteHTML.IntrinsicElements['p'] & {
-	node?: HTMLParagraphElement | undefined | null;
-};
