--- packages/dialog/svelte/lib/DialogRoot.ts
@@ -1,4 +1,0 @@
-export interface DialogRootProps {
-	open?: boolean;
-	initialOpen?: boolean;
-}
