--- packages/dialog/svelte/lib/DialogTitle.ts
@@ -1,3 +1,0 @@
-export type DialogTitleProps = svelteHTML.IntrinsicElements['h1'] & {
-	node?: HTMLHeadingElement | undefined | null;
-};
