--- packages/dialog/svelte/lib/DialogTrigger.ts
@@ -1,3 +1,0 @@
-export type DialogTriggerProps = svelteHTML.IntrinsicElements['button'] & {
-	node?: HTMLButtonElement | undefined | null;
-};
